chore(commands): move files to better fitting folders
